refactor: first drop of refactored cmake aiming at separating librarires
