+++ opensfm/src/dense/src/depthmap_test.cc
@@ -1,11 +1,11 @@
+#include <dense/depthmap.h>
 #include <opencv2/core/core.hpp>
 #include <opencv2/calib3d/calib3d.hpp>
 #include <gtest/gtest.h>
-#include "depthmap.h"
 
 namespace {
 
-using namespace csfm;
+using namespace dense;
 
 
 TEST(PlaneInducedHomography, RandomPoint) {
