+++ opensfm/src/foundation/src/logger.cc
@@ -1,0 +1,9 @@
+#include <glog/logging.h>
+#include <foundation/types.h>
+
+class GLogInitializationWrapper {
+ public:
+  GLogInitializationWrapper() { google::InitGoogleLogging("opensfm"); }
+};
+
+static GLogInitializationWrapper initializer;
