+++ opensfm/src/geometry/python/pybind.cc
@@ -3,19 +3,16 @@
 #include <pybind11/eigen.h>
 #include <glog/logging.h>
 
-#include "essential.h"
-#include "pose.h"
-#include "triangulation.h"
+#include <geometry/essential.h>
+#include <geometry/pose.h>
+#include <geometry/triangulation.h>
+#include <foundation/types.h>
 
-namespace py = pybind11;
 
-
-PYBIND11_MODULE(csfm, m) {
-  google::InitGoogleLogging("csfm");
-
-  m.def("triangulate_bearings_dlt", csfm::TriangulateBearingsDLT);
-  m.def("triangulate_bearings_midpoint", csfm::TriangulateBearingsMidpoint);
-  m.def("essential_five_points", csfm::EssentialFivePoints);
-  m.def("essential_n_points", csfm::EssentialNPoints);
-  m.def("relative_pose_from_essential", csfm::RelativePoseFromEssential);
+PYBIND11_MODULE(pygeometry, m) {
+  m.def("triangulate_bearings_dlt", geometry::TriangulateBearingsDLT);
+  m.def("triangulate_bearings_midpoint", geometry::TriangulateBearingsMidpoint);
+  m.def("essential_five_points", geometry::EssentialFivePoints);
+  m.def("essential_n_points", geometry::EssentialNPoints);
+  m.def("relative_pose_from_essential", geometry::RelativePoseFromEssential);
 }
