+++ opensfm/src/geometry/src/pose.cc
@@ -1,0 +1,18 @@
+#include <geometry/pose.h>
+
+namespace geometry {
+Eigen::Matrix<double, 3, 4> RelativePoseFromEssential(
+    const Eigen::Matrix3d& essential,
+    const Eigen::Matrix<double, -1, 3> &x1,
+    const Eigen::Matrix<double, -1, 3> &x2) {
+  if((x1.cols() != x2.cols()) || (x1.rows() != x2.rows())){
+    throw std::runtime_error("Features matrices have different sizes.");
+  }
+  std::vector<std::pair<Eigen::Vector3d, Eigen::Vector3d>> samples(x1.rows());
+  for (int i = 0; i < x1.rows(); ++i) {
+    samples[i].first = x1.row(i);
+    samples[i].second = x2.row(i);
+  }
+  return ::RelativePoseFromEssential(essential, samples.begin(), samples.end());
+}
+}
