+++ opensfm/src/robust/python/pybind.cc
@@ -3,12 +3,10 @@
 #include <pybind11/eigen.h>
 #include <glog/logging.h>
 
-#include "robust_estimator.h"
-#include "scorer.h"
-#include "instanciations.h"
-
-
-namespace py = pybind11;
+#include <foundation/types.h>
+#include <robust/robust_estimator.h>
+#include <robust/scorer.h>
+#include <robust/instanciations.h>
 
 
 template <class T>
@@ -22,9 +20,7 @@ void AddScoreType(py::module& m, const std::string& name) {
     ;
   }
 
-PYBIND11_MODULE(csfm, m) {
-  google::InitGoogleLogging("csfm");
-
+PYBIND11_MODULE(pyrobust, m) {
   py::class_<RobustEstimatorParams>(m, "RobustEstimatorParams")
     .def(py::init())
     .def_readwrite("iterations", &RobustEstimatorParams::iterations)
@@ -33,9 +29,9 @@ PYBIND11_MODULE(csfm, m) {
     .def_readwrite("use_iteration_reduction", &RobustEstimatorParams::use_iteration_reduction)
   ;
 
-  m.def("ransac_line", csfm::RANSACLine);
-  m.def("ransac_essential", csfm::RANSACEssential);
-  m.def("ransac_relative_pose", csfm::RANSACRelativePose);
+  m.def("ransac_line", robust::RANSACLine);
+  m.def("ransac_essential", robust::RANSACEssential);
+  m.def("ransac_relative_pose", robust::RANSACRelativePose);
 
   py::enum_<RansacType>(m, "RansacType")
       .value("RANSAC", RansacType::RANSAC)
@@ -44,6 +40,6 @@ PYBIND11_MODULE(csfm, m) {
       .export_values()
     ;
   AddScoreType<Line::MODEL>(m, "Line");
-  AddScoreType<csfm::EssentialMatrixModel::MODEL>(m, "EssentialMatrix");
+  AddScoreType<robust::EssentialMatrixModel::MODEL>(m, "EssentialMatrix");
   AddScoreType<Eigen::Matrix<double, 3, 4>>(m, "RelativePose");
 }
