+++ setup.py
@@ -67,7 +67,12 @@ setuptools.setup(
     ],
     package_data={
         'opensfm': [
-            'csfm.*',
+            'pybundle.*',
+            'pygeometry.*',
+            'pyrobust.*',
+            'pyfeatures.*',
+            'pydense.*',
+            'pyfoundation.*',
             'data/sensor_data.json',
             'data/bow/bow_hahog_root_uchar_10000.npz',
             'data/bow/bow_hahog_root_uchar_64.npz',
